requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
atari-py==0.2.6
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
Brlapi==0.6.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudpickle==1.6.0
command-not-found==0.3
config==0.5.0.post0
constantly==15.1.0
cryptography==3.2.1
cupshelpers==1.0
cycler==0.10.0
dateparser==1.0.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
future==0.18.2
gym==0.17.3
httplib2==0.9.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
joblib==0.17.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
krakenex==2.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.6.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.3.2
multitasking==0.0.9
netifaces==0.10.4
numpy==1.19.2
oauth==1.0.1
olefile==0.45.1
opencv-python==4.4.0.44
pandas==1.1.3
pexpect==4.2.1
Pillow==7.2.0
protobuf==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyglet==1.5.0
pygobject==3.26.1
PyHamcrest==2.0.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-binance==0.7.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
regex==2020.10.28
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
scipy==1.5.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.13.2
six==1.15.0
ssh-import-id==5.7
stable-baselines==2.10.1
system-service==0.3
systemd-python==234
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
wadllib==1.3.2
wget==3.2
xkit==0.0.0
yfinance==0.1.55
zope.interface==5.1.2
